Add QoL improvements - Back Btn and Barline Disp
Features:
* New Back Button when trying to create a new map. Clicking
it will bring the user back to the main menu, where a new
audio file could be loaded.
* If no new audio file was loaded, then the existing values
in the fields for creating a new map would not be reset.
* If a new audio file was loaded, only the map file name
would be reset.

* Change the default number of toggle options for down, left
and right arrow keys to 2. (Down from 3). By default, pressing
those arrow keys would toggle the translucent barlines on/off.

* This means that solid barlines are no longer shown by default.
However, they could be turned back on by changing the corresponding
`includeSolidBarLine` option to `true` in main.js.